+++ OctoberProject/MainController/MainCells/Struct/MainCollectionItems.swift
