+++ OctoberProject/MainController/MainCells/Struct/PetMediaCollectionItems.swift
